[FILTER][AMI|VM] Refactor filter custom component
