--- src/components/custom/fields/index.ts
@@ -1,1 +1,0 @@
-export * from "./filter-item-checkbox/filter-item-checkbox";
