+++ src/components/custom/filter/index.ts
@@ -1,2 +1,3 @@
 export * from "./filter/filter";
 export * from "./filter-item/filter-item";
+export * from "./filter-item-checkbox/filter-item-checkbox";
